1.1.0-dev.7
Added
-   Project now support typescript!

### Changed
-   Some README.md rework

### Removed
-   Some unused code disappeared.
