+++ CHANGELOG.md
@@ -10,14 +10,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 -   Firmware Revision configuration option (optional) [#211](https://github.com/NRCHKB/node-red-contrib-homekit-bridged/pull/211)
 -   Hardware and Software Revision configuration option (optional)
+-   Project now support typescript!
 
 ### Fixed
 -   Error status is not passed to callback when "No response" was triggered [#227](https://github.com/NRCHKB/node-red-contrib-homekit-bridged/pull/227) also discussed in [#185](https://github.com/NRCHKB/node-red-contrib-homekit-bridged/pull/185)
 
 ### Changed
+-   Some README.md rework
 -   Updated hap-nodejs to 0.7.3
 -   Updated dependencies to latest versions
 -   Moved lint-staged config to main level and added minimum nodejs version
+
+### Removed
+-   Some unused code disappeared.
 
 ## [1.0.4] - 2020.03.03
