+++ package.json
@@ -1,122 +1,131 @@
 {
-    "name": "node-red-contrib-homekit-bridged",
-    "version": "1.1.0-dev.6",
-    "description": "Node-RED nodes to simulate Apple HomeKit devices.",
-    "main": "homekit.js",
-    "scripts": {
-        "test": "mocha --exit",
-        "fix": "eslint --fix \"**/*.js\"",
-        "lint": "eslint \"**/*.js\""
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/NRCHKB/node-red-contrib-homekit-bridged"
-    },
-    "keywords": [
-        "homekit",
-        "hap-nodejs",
-        "node-red",
-        "iot",
-        "homebridge",
-        "home",
-        "NRCHKB"
-    ],
-    "node-red": {
-        "nodes": {
-            "homekit": "homekit.js"
-        }
-    },
-    "author": {
-        "name": "Shaquu",
-        "email": "shaquu.github@gmail.com"
-    },
-    "contributors": [
-        "Marius Schmeding",
-        "Oliver Rahner"
-    ],
-    "license": "Apache-2.0",
-    "bugs": {
-        "url": "https://github.com/NRCHKB/node-red-contrib-homekit-bridged/issues"
-    },
-    "homepage": "https://github.com/NRCHKB/node-red-contrib-homekit-bridged#readme",
-    "dependencies": {
-        "debug": "^4.1.1",
-        "hap-nodejs": "^0.7.3"
-    },
-    "devDependencies": {
-        "eslint": "^7.0.0",
-        "eslint-config-prettier": "^6.11.0",
-        "eslint-plugin-prettier": "^3.1.3",
-        "husky": "^4.2.5",
-        "lint-staged": "^10.2.4",
-        "mocha": "^7.1.2",
-        "node-red": "^1.0.6",
-        "node-red-node-test-helper": "^0.2.3",
-        "prettier": "^2.0.5"
-    },
-    "eslintConfig": {
-        "env": {
-            "browser": true,
-            "commonjs": true,
-            "es6": true,
-            "node": true
-        },
-        "extends": [
-            "prettier",
-            "eslint:recommended"
-        ],
-        "plugins": [
-            "prettier"
-        ],
-        "globals": {
-            "Atomics": "readonly",
-            "SharedArrayBuffer": "readonly"
-        },
-        "parserOptions": {
-            "ecmaVersion": 2018
-        },
-        "rules": {
-            "prettier/prettier": "off",
-            "indent": [
-                "error",
-                4
-            ],
-            "linebreak-style": [
-                "error",
-                "unix"
-            ],
-            "quotes": [
-                "error",
-                "single"
-            ],
-            "semi": [
-                "error",
-                "never"
-            ],
-            "no-prototype-builtins": "off"
-        }
-    },
-    "prettier": {
-        "tabWidth": 4,
-        "useTabs": false,
-        "semi": false,
-        "singleQuote": true,
-        "trailingComma": "es5",
-        "endOfLine": "lf"
-    },
-    "husky": {
-        "hooks": {
-            "pre-commit": "lint-staged && yarn lint",
-            "pre-push": "yarn lint"
-        }
-    },
-    "lint-staged": {
-        "*.js": [
-            "eslint --fix",
-            "git add"
-        ]
-    },
-    "engines": {
-        "node": ">=10"
+  "name": "node-red-contrib-homekit-bridged",
+  "version": "1.1.0-dev.7",
+  "description": "Node-RED nodes to simulate Apple HomeKit devices.",
+  "main": "build/homekit.js",
+  "scripts": {
+    "build": "yarn clean && tsc",
+    "clean": "rimraf build/lib build/**/*.js",
+    "test": "mocha -r ts-node/register './src/**/*.test.js' --exit",
+    "fix": "eslint --fix \"src/**/*.js\"",
+    "lint": "eslint \"src/**/*.js\""
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/NRCHKB/node-red-contrib-homekit-bridged"
+  },
+  "keywords": [
+    "homekit",
+    "hap-nodejs",
+    "node-red",
+    "iot",
+    "homebridge",
+    "home",
+    "NRCHKB"
+  ],
+  "node-red": {
+    "nodes": {
+      "homekit": "build/homekit.js"
     }
+  },
+  "author": {
+    "name": "Shaquu",
+    "email": "shaquu.github@gmail.com"
+  },
+  "contributors": [
+    "Marius Schmeding",
+    "Oliver Rahner"
+  ],
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/NRCHKB/node-red-contrib-homekit-bridged/issues"
+  },
+  "homepage": "https://github.com/NRCHKB/node-red-contrib-homekit-bridged#readme",
+  "dependencies": {
+    "debug": "^4.1.1",
+    "hap-nodejs": "^0.7.3"
+  },
+  "devDependencies": {
+    "@types/debug": "^4.1.5",
+    "@types/express": "^4.17.6",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "^10.17.24",
+    "@types/node-red": "^0.20.1",
+    "eslint": "^7.1.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.3",
+    "husky": "^4.2.5",
+    "lint-staged": "^10.2.6",
+    "mocha": "^7.2.0",
+    "node-red": "^1.0.6",
+    "node-red-node-test-helper": "^0.2.4",
+    "prettier": "^2.0.5",
+    "ts-node": "^8.10.1",
+    "typescript": "^3.9.3"
+  },
+  "eslintConfig": {
+    "env": {
+      "browser": true,
+      "commonjs": true,
+      "es6": true,
+      "node": true
+    },
+    "extends": [
+      "prettier",
+      "eslint:recommended"
+    ],
+    "plugins": [
+      "prettier"
+    ],
+    "globals": {
+      "Atomics": "readonly",
+      "SharedArrayBuffer": "readonly"
+    },
+    "parserOptions": {
+      "ecmaVersion": 2018
+    },
+    "rules": {
+      "prettier/prettier": "off",
+      "indent": [
+        "error",
+        4
+      ],
+      "linebreak-style": [
+        "error",
+        "unix"
+      ],
+      "quotes": [
+        "error",
+        "single"
+      ],
+      "semi": [
+        "error",
+        "never"
+      ],
+      "no-prototype-builtins": "off"
+    }
+  },
+  "prettier": {
+    "tabWidth": 4,
+    "useTabs": false,
+    "semi": false,
+    "singleQuote": true,
+    "trailingComma": "es5",
+    "endOfLine": "lf"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged && yarn lint",
+      "pre-push": "yarn lint"
+    }
+  },
+  "lint-staged": {
+    "*.js": [
+      "eslint --fix",
+      "git add"
+    ]
+  },
+  "engines": {
+    "node": ">=10"
+  }
 }
