+++ src/lib/HAPBridgeNode.js
