+++ src/lib/HAPServiceNode.js
