+++ src/lib/cameraSource/index.js
