+++ src/lib/utils/BridgeUtils.js
