+++ src/lib/utils/CharacteristicUtils.js
